Create a topic listing stub
Added real topic listing to the board topics page. It’s just a stub,
since the counters don’t work and the topic subjects are not
links. They’re real topics belonging to the board though. There is a
link button to create a new topic in the respective board.